+++ CHANGELOG.md
@@ -4,5 +4,14 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.0] - 2024-01-10
+### Added
+- Settings to change the note (base) size and scaling options.
+
+### Changed
+- Notes and the font size now scale with the size of the canvas. They are scaled so that roughly eight notes can be placed along the shorter side of the map. This scale can be changed so that either 16 (with small notes) or six (with large notes) notes can be placed along the shorter side of the map. This scaling can be disabled, so that the notes always have a size of 512 px (or 256 px or 768 px, respectively).
+- The Sticky Note buttons has been moved directly below the text drawing button.
+- The setting labels have been clarified.
+
 ## [1.0.0] - 2024-01-09
 - Initial Release